requirements.txt
asgiref==3.8.1
boto3==1.26.37
botocore==1.29.165
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
dj-database-url==3.0.0
Django==4.2.13
django-phonenumber-field==7.3.0
django-social-share==2.3.0
django-storages==1.14.4
idna==3.10
jmespath==1.0.1
phonenumbers==8.13.38
pillow==10.4.0
pycoingecko==2.2.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
qrcode==7.4.2
requests==2.32.3
s3transfer==0.6.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.20
whitenoise==6.9.0
